requirements.txt
anyio==3.6.2
APScheduler==3.6.3
cachetools==4.2.2
certifi==2022.12.7
dill==0.3.6
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
mysql-connector-python==8.0.32
numpy==1.24.2
pandas==1.5.3
protobuf==3.20.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-telegram-bot==13.15
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
tornado==6.1
tzdata==2022.7
tzlocal==4.2
